Fixed `ArgParser` lifetimes
Now stores a mutable reference to the underlying zval. Instead of
passing the execution data to the arg parser, a vector of args is now
passed to prevent a shared lifetime issue with the `$this` object.

+++ src/errors.rs
@@ -21,7 +21,7 @@ pub enum Error {
     /// The enum carries two integers - the first representing the minimum
     /// number of arguments expected, and the second representing the number of
     /// arguments that were received.
-    IncorrectArguments(u32, u32),
+    IncorrectArguments(usize, usize),
     /// There was an error converting a Zval into a primitive type.
     ///
     /// The enum carries the data type of the Zval.
